Move SendCardPopup to molecules in CardsCreator #371
