+++ WMIAdventure/frontend/src/js/pages/creative/CardsCreator/molecules/SendCardPopup/styled-components/ContentContainer.js
